Yossi/map generation (#6)
* - Add a class to generate map.
- The class can generate Intersection nodes.
- Now creating the roude edges
- After that, i need to build the cells according to the edges and nodes

* Continued work on the Road buildings

* Complited the FillRoads function.
It will create a road like this:
each cell will have multiple 'connectedPoints' which indicates from where to where the road gose to.
This should help creating movment paths for persons later.
So when we want to add a road to a cell, we add two points. 1 in the certer, and 1 in the direction of cell building. Also this should be done in the next cell on the oposit direction.

* Fixed a compilation error in City.cs

* Renamed city to map

* Map Generation works great now.
Step 1 is complete

* Add the free tiles too

Co-authored-by: Yossi Alfarsi <yosal@microsoft.com>